+++ brane-cli/src/packages.rs
@@ -373,7 +373,7 @@ pub async fn remove(
 
             // Remove that image from the Docker daemon
             let image: Image = Image::new(&package_info.name, Some(format!("{}", package_info.version)), Some(digest));
-            if let Err(err) = docker::remove_image(docker_opts, &image).await {
+            if let Err(err) = docker::remove_image(&docker_opts, &image).await {
                 return Err(PackageError::DockerRemoveError{ image: Box::new(image), err });
             }
 
@@ -468,7 +468,7 @@ pub async fn remove(
 
             // Remove that image from the Docker daemon
             let image: Image = Image::new(&package_info.name, Some(format!("{}", package_info.version)), Some(digest));
-            if let Err(err) = docker::remove_image(docker_opts, &image).await {
+            if let Err(err) = docker::remove_image(&docker_opts, &image).await {
                 return Err(PackageError::DockerRemoveError{ image: Box::new(image), err });
             }
         }